Finished ability to read text from file to generate address book.
Instead of hardcoding the remote addresses into the source code, we can
now load a a document containing the name and address of each Xbee on a
separate line to populate the addressbook. This is important because it
will eventually enable us to deply an application and still be able to
change the internal settings via save file.